NumbersGameLab_Programming1 (zipped)
The purpose of this lab was to make a game about reducing a starting integer to 0 with user inputs of either 1 or 2.
I got it working fairly quickly, so I used the extra time I had to make it goofy.

EXAMPLE OUTPUT:

Welcome to JackBlack!

It's just like the classic blackjack you know and love!
Except, you both have to share a starting hand that already adds up to 21 or less.
Also, you can pick any card! But only as long as it's a 1 or 2. This subtracts from the hand.
The player who brings it to zero or below wins!

Good luck!

Enter player names to begin

Player One: Alejandro
Player Two: Also Alejandro

Your starting hand is: 18

Alejandro, it's your turn!
Enter '1' or '2' :  2

Current hand is: 16

Also Alejandro, it's your turn!
Enter '1' or '2' :  3

Not a 1 or a 2! Do it again!

Enter '1' or '2' :  e

Not even an integer! Did you read the rules?
Not a 1 or a 2! Do it again!

Enter '1' or '2' :  1

Current hand is: 15

(skipping ahead...)

Current hand is: 2

Alejandro, it's your turn!
Enter '1' or '2' :  2

And we're down to 0!
Alejandro won! Play again? [y/n + enter]